refactor: item, inventory 관련 파일 script 도메인 폴더 내로 이동
